# Conflicts:
#	MBKM.BusinessLogic/MBKM.Common/Interfaces/RepoInterfaces/MBKMRepoInterfaces/INilaiKuliahRepository.cs
#	MBKM.BusinessLogic/MBKM.Services/MBKMServices/NilaiKuliahService.cs
#	MBKM.Data/MBKM.Entities/MBKM.Entities.csproj
#	MBKM.Data/MBKM.Repository/Repositories/MBKMRepository/NilaiKuliahRepository.cs
#	MBKM.UI/MBKM.Presentation/MBKM.Presentation.csproj

+++ MBKM.Data/MBKM.Entities/ViewModel/VMSubBobot .cs
@@ -10,6 +10,6 @@ namespace MBKM.Entities.ViewModel
     {
         public string Nama { get; set; }
         public string Jenis { get; set; }
-        public float Persentase { get; set; }
+        public double Persentase { get; set; }
     }
 }
